+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `PixelMap::map_size` to store a logical size for the map that is not restrained by the requirements
   of the underlying node implementation, which requires dimensions that are square and powers of two.
 * Refactor shapes to be represented in unsigned space.
+* Add `PixelMap::contour` method to obtain a list of line segments that contour a shape.
 
 ## v0.2.0